FlarmTrafficWindow: GetZoomDistance() returns integer
This may take advantage of a "fixed" optimization, in the division
operator.

+++ src/Gauge/FlarmTrafficWindow.hpp
@@ -113,7 +113,7 @@ public:
   void SelectNearTarget(int x, int y);
 
 protected:
-  static double GetZoomDistance(unsigned zoom);
+  static unsigned GetZoomDistance(unsigned zoom);
 
   void GetZoomDistanceString(TCHAR* str1, TCHAR* str2, unsigned size) const;
   double RangeScale(double d) const;
